feat: migrate RoutineSection logic to Home page and remove RoutineSection component
